refactor(project-structure): organize folders to follow React standard conventions
