+++ src/components/Dashboard/Dashboard.jsx
@@ -6,8 +6,8 @@ import { DashboardItem } from "./DashBoardItem";
 import { SiBookstack } from "react-icons/si";
 import { RxAvatar } from "react-icons/rx";
 import { BsThreeDots } from "react-icons/bs";
-import { UseGetLoggedInUserApi } from "../../Queries/UseGetLoggedInUserApi";
-import { Spinner } from "../../uiutils/Spinner";
+import { UseGetLoggedInUserApi } from "../../hooks/queries/UseGetLoggedInUserApi";
+import { Spinner } from "../Spinner";
 import { CiLogout } from "react-icons/ci";
 import { removeToken } from "../../utils/Token";
 import { useNavigate } from "react-router-dom";
