+++ src/components/Students/EditStudentForm.jsx
@@ -2,15 +2,15 @@
 import { useState } from "react";
 
 import { useForm } from "react-hook-form";
-import { Spinner } from "../../uiutils/Spinner";
+import { Spinner } from "../Spinner";
 import { uploadToCloudinary } from "../../utils/uploadToCloudinary";
 
-import { RadioInput } from "../../uiutils/RadioInput";
-import { FileInput } from "../../uiutils/FileInput";
+import { RadioInput } from "../RadioInput";
+import { FileInput } from "../FileInput";
 import { HiXMark } from "react-icons/hi2";
-import { SubmitButton } from "../../uiutils/SubmitButton";
-import { Input } from "../../uiutils/Input";
-import { useUpdateStudent } from "../../Queries/Student/useUpdateStudent";
+import { SubmitButton } from "../SubmitButton";
+import { Input } from "../Input";
+import { useUpdateStudent } from "../../hooks/queries/Student/useUpdateStudent";
 import { UPDATE_STUDENT } from "../../utils/Routes";
