+++ src/hooks/queries/Department/useDeleteDept.js
