+++ src/hooks/queries/Student/useAddStudent.js
@@ -1,5 +1,5 @@
 import { useMutation } from "@tanstack/react-query"
-import { postApi } from "../../ApisCalls/PostApi"
+import { postApi } from "../../api/PostApi"
 import toast from "react-hot-toast"
 
 export const useAddStudent=()=>{
