+++ src/hooks/queries/Teacher/useGetAllTeachers.js
@@ -2,7 +2,7 @@ import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { VIEW_ALL_TEACHERS } from "../../utils/Routes";
 import toast from "react-hot-toast";
 import { useSearchParams } from "react-router-dom";
-import { getAllApi } from "../../ApisCalls/getAllApi";
+import { getAllApi } from "../../api/getAllApi";
 export const useGetAllTeachers=()=>{
     const [searchParam] = useSearchParams();
     // console.log("SearchParam", searchParam);
